+++ src/tests/TestAnchoredCurve/anchoredcurve.cpp
@@ -3,8 +3,8 @@
 #include <algorithm>
 #include <utility>
 
-template class AnchoredCurve<int, int>;
-template class Knot<int, int>;
+template class AnchoredCurve<int, double>;
+template class Knot<int, double>;
 
 template <typename TPos, typename TValue>
 Knot<TPos, TValue>::Knot() : position(TPos()), value(TValue()), slope(0.0) {
